Deleted a component
