+++ src/App.css
@@ -55,31 +55,6 @@ header{
   animation: rotation infinite 4s linear;
 }
 
-/* .earth-image{
-  transform-style: preserve-3d;
-  transform: rotateX(-40deg) rotateY(32deg);
-perspective: 1000px;
-}
-.earth-image img{
-    width:882px;
-      height: 561px;
-      box-shadow: rem(-20) rem(20) rem(80) 0 rgba(0, 0, 0, 0.15),
-                  rem(-20) rem(20) rem(80) 0  rgba(0, 0, 0, 0.15) ;
-   transform: rotateX( 90deg)  translateZ(1em);
-} */
-
-
-/* @keyframes rotation{
-  
-  0% {
-    transform: rotateY(0deg);
-  }
-  100% {
-    transform: rotateY(359deg);
-  }
-} */
- 
- 
 .heading-container{
   display:flex;
   color:white;
@@ -117,7 +92,7 @@ h3{
 
 .link-image{
   margin-left: 31%;
-  margin-top: -19px;
+  margin-top: -11px;
   width: 115px;
   position:relative;
   transition:.3s linear;
@@ -166,8 +141,8 @@ nav a{
   position: absolute;
   top: 4px;
   right: -7px;
-  width: 24px;
-  height: 24px;
+  width: 31px;
+  height: 31px;
   border:2px solid black;
   font-style:bold;
   transition:.3s linear
@@ -299,11 +274,12 @@ img{
   padding:5px 10px;
   background-color: #28282b;
   flex-grow:1;
-  transition:.4s ease;
+  transition:.6s linear;
   text-align:center;
   background-image: linear-gradient(#28282b,#57575b);
   border-radius:5px;
   font-family: 'Varela Round', sans-serif;
+
 }
 
 .form-buttons:hover{
