+++ src/Footer.js
@@ -2,7 +2,7 @@ const Footer = () =>{
     return(
         <footer>
             <div className="wrapper">
-                <p>Designed and Built by Joe Gencarelli using the NEWS API</p>
+                <p>Designed and Built by Joe Gencarelli using the Currents API</p>
                 <p>Created at <a href="https://junocollege.com"><span className="footer-span">Juno College</span></a></p>
             </div>
         </footer>
